Corrected stupid mistake in skin model
git-svn-id: http://benbelly.unfuddle.com/svn/benbelly_ms/epenthesis@91 c3643d77-5941-4ca3-9a90-978904f1a040

+++ SkinModel.cpp
@@ -7,7 +7,7 @@
 
 std::pair<int, double *> loadSkinModel() {
     FILE *in = fopen( "skinmodel.model", "rb" );
-    int step = 0;
+    double step = 0;
     fread( &step, sizeof(double), 1, in );
 
     int l = 256 / step;
